+++ cli/src/connect/__test__/e2e/e2e_parse_command/custom_parser_error_no_include/parser/custom_parser.js
@@ -1,0 +1,15 @@
+let data = ''
+
+process.stdin.on('data', function (chunk) {
+  data += chunk
+})
+
+process.stdin.on('end', function () {
+  const parsed = JSON.parse(data)
+  const result = {
+    docs: [],
+    messages: [{ level: 'ERROR', message: 'Error from parser!' }],
+  }
+
+  process.stdout.write(JSON.stringify(result))
+})