+++ cli/src/connect/__test__/e2e/e2e_parse_command/custom_parser_invalid_response/parser/custom_parser.js
@@ -1,0 +1,22 @@
+let data = ''
+
+process.stdin.on('data', function (chunk) {
+  data += chunk
+})
+
+process.stdin.on('end', function () {
+  const parsed = JSON.parse(data)
+  // Fake result which passes back the input, formatted to match the expected
+  // schema, so we can validate it was passed in correctly
+  const result = {
+    docs: parsed.paths.map((path) => ({
+      test: 123,
+    })),
+    messages: [
+      { level: 'DEBUG', message: 'Debug message from parser!' },
+      { level: 'INFO', message: 'Success from parser!' },
+    ],
+  }
+
+  process.stdout.write(JSON.stringify(result))
+})